cli/agents/meeting_stratergy_agent/pyproject.toml
[project]
name = "meeting-strategy-agent"
version = "0.1.0"
description = "Meeting Strategy Agent - develops talking points and strategic angles for meetings"
requires-python = ">=3.12"
dependencies = [
    "genai-protocol>=1.0.9",
    "python-dotenv>=1.1.0",
    "google-generativeai>=0.8.0",
]
